+++ lib/models/dount_chart/chart_data.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class ChartData {
+  final double value;
+  final Color color;
+
+  ChartData({required this.value, required this.color});
+}
